+++ tests/Flare.Components.Tests/TestHelpers.cs
@@ -211,6 +211,7 @@ public class FlareTestContext : BunitContext
         Services.AddScoped<Flare.Components.Services.IDialogDragJsService, Flare.Components.Services.DialogDragJsService>();
         Services.AddScoped<Flare.Components.Services.IDragDropJsService, Flare.Components.Services.DragDropJsService>();
         Services.AddScoped<Flare.Components.Services.IChartMotionJsService, Flare.Components.Services.ChartMotionJsService>();
+        Services.AddScoped<Flare.Components.Services.IProgressGeometryJsService, Flare.Components.Services.ProgressGeometryJsService>();
         Services.AddScoped<Flare.Components.IBrowserViewportService, Flare.Components.Services.BrowserViewportService>();
         Services.AddScoped<Flare.Components.Services.IColorCanvasJsService, Flare.Components.Services.ColorCanvasJsService>();
         Services.AddScoped<Flare.Components.Services.IHighlightJsService, Flare.Components.Services.HighlightJsService>();